backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-cloud-aiplatform==1.145.0
google-cloud-alloydb-connector[asyncpg]==1.6.0
google-adk==1.0.0
SQLAlchemy==2.0.35
pgvector==0.3.2
python-multipart==0.0.22
pydantic==2.10.4
pypdf==6.9.2
python-docx==1.1.2
PyYAML==6.0.2
sse-starlette==2.2.1
python-dotenv==1.0.1
aiofiles==24.1.0
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
